fix: exit code handling (#5)
The problem with using `exit` instead of `return` is it hides the fact
that your error handling isn't working correctly. Especially in the
context of `set -e`.

If you do the hard work of making your functions return a non-zero exit
code all the way up the chain back to the entrypoint of your script, and
your script exits the way you expect, then everything's gravy.

However if you just `exit` from anywhere in your script, then 99% of the
time you'll crash your script as expected, but you might be hiding the
fact that _other_ commands that have non-zero exit codes are _not_
crashing your script as you expect.

In this PR, I replace a few `exit` statements with `return`, and in the
process revealed a bug that allowed `apply` to fail ONLY if you
explicitly call `panic` or `exit 1`. Before this change, `apply` would
not fail if, for example, you ran the `false` command.

- **chore: make __BLARG_MAIN__ use return instead of exit**
- **fix `apply` return handling**
- **more useful output after an apply failure**